VisualStringDistances

VisualStringDistances.Glyph

VisualStringDistances.Glyph

VisualStringDistances.GlyphCoordinates

VisualStringDistances.glyph!

VisualStringDistances.Glyph — Type

Glyph <: AbstractArray{Bool,2}

Holds the bitmap associated to a Unifont glyph in a packed format.

VisualStringDistances.Glyph — Method

Glyph(s::String) --> Glyph

Construct a Glyph from a string.

Examples

julia> Glyph("abc")
------------------------
------------------------
------------------------
---------#--------------
---------#--------------
---------#--------------
--####---#-###----####--
-#----#--##---#--#----#-
------#--#----#--#------
--#####--#----#--#------
-#----#--#----#--#------
-#----#--#----#--#------
-#---##--##---#--#----#-
--###-#--#-###----####--
------------------------
------------------------

VisualStringDistances.GlyphCoordinates — Type

GlyphCoordinates{T} <: AbstractVector{T}

A sparse representation of a Glyph.

VisualStringDistances.glyph! — Method

glyph!(v::Vector{UInt8}) -> Glyph

Creates a Glyph for a vector of bytes, assuming the vector represents a single Unifont character. Modifies v and may share its memory.